Refactored Logging environment (cleaned up)
